Change IUserAnswer.EvaluateScore return type to float and implement in all SurveyPages

+++ MobileDataCollection.Survey/MobileDataCollection.Survey/Models/AnswerIntrospectionPage.cs
@@ -38,7 +38,7 @@ namespace MobileDataCollection.Survey.Models
             SelectedAnswer = selectedanswer;
         }
 
-        public int EvaluateScore()
+        public float EvaluateScore()
         {
             throw new NotSupportedException();
         }
